@@ -2,7 +2,8 @@ package com.icechn.videorecorder.tools;
 
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import java.util.concurrent.Executor;
 
